Adding cwd (current working directory) support to Posix-spawned processes. You can now set the current working directory for a process before it starts. Still needs formal tests.
Unfortunately, because posix_spawnp does not take any argument to change the child's current working directory, we have to change the parent process' directory so that it gets inherited. Additionally, since the current directory is a process-wide setting and not per-thread, this opens up a small window for race conditions since the chdir might conflict with other directory changes in other threads. If this is an issue, the guidance would be to stick with fork/exec.