Spec environment unset in parent is unset in parent when set in grandparent
MSP-11414

"can unset env vars" checks that with CHILDPROCESS_UNSET=1 in
grandparent (the shell) and parent, the child can unset it by explicitly setting
`process.environment['CHILDPROCESS_UNSET'] = nil`; however, this doesn't
cover that environment variables being unset in the parent, that were
set in the grandparent are unset in the child. "does not see env var
unset in parent" checks this, and it currently fails for JRuby.