add JRuby 9.2 and Java 11 to the ci mix
The `JAVA_OPTS` environment variable is used by JRuby to launch the JVM with
appropriate options; by adding specific `--add-opens` flags, we are able to
enable the specific reflective access that is necessary to run in Javas >= 9.

For ci, we set `JAVA_OPTS` with the contents of `JAVA_OPTS_FOR_SPECS` inside
the `before_script` in order to avoid invoking all jruby JVMs with the flags.